Extract ActionButton to new class IconButton MANY BUTTONS WITH ICONS
